+++ internal/script/payout.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/NodeFactoryIo/vedran/internal/controllers"
-	"github.com/NodeFactoryIo/vedran/internal/models"
 	"github.com/NodeFactoryIo/vedran/internal/payout"
 	"net/http"
 	"net/url"
@@ -13,20 +12,15 @@ import (
 var statsEndpoint, _ = url.Parse("/api/v1/stats")
 
 func ExecutePayout(secret string, totalReward float64, loadbalancerUrl *url.URL) ([]*payout.TransactionDetails, error) {
-	stats, err := fetchStatsFromEndpoint(loadbalancerUrl.ResolveReference(statsEndpoint))
+	response, err := fetchStatsFromEndpoint(loadbalancerUrl.ResolveReference(statsEndpoint))
 	if err != nil {
 		return nil, fmt.Errorf("unable to fetch stats from loadbalancer, %v", err)
 	}
 
-	// calculate distribution
-	nodeStatsDetails := make(map[string]models.NodeStatsDetails, len(stats.Stats))
-	for nodeId, nodeStats := range stats.Stats {
-		nodeStatsDetails[nodeId] = nodeStats.Stats
-	}
 	distributionByNode := payout.CalculatePayoutDistributionByNode(
-		nodeStatsDetails,
+		response.Stats,
 		totalReward,
-		float64(stats.Fee),
+		float64(response.Fee),
 	)
 
 	return payout.ExecuteAllPayoutTransactions(